+++ guillotina/tests/test_api.py
@@ -1245,7 +1245,7 @@ class ITestSchema(Interface):
 
     object_a = schema.Object(IObjectA, required=False)
     list_object_a = PatchField(schema.List(value_type=schema.Object(IObjectA), required=False))
-    time_ = schema.Time(min="9:00:00", max="11:00:00", required=False)
+    time_ = schema.Time(min=time(9, 0, 0), max=time(12, 0, 0), required=False)
 
 
 @contenttype(type_name="TestSchema", schema=ITestSchema)
@@ -1320,29 +1320,49 @@ async def test_field_values_list_bucket(container_requester):
         )
         assert status == 410
 
+
 async def test_time_field_validation(container_requester):
     async with container_requester as requester:
-        __import__("pdb").set_trace()
         resp, status = await requester(
-            "POST", "/db/guillotina/", data=json.dumps({"@type": "TestSchema", "time_": "16:40:09", "id": "foo_item"})
+            "POST",
+            "/db/guillotina/",
+            data=json.dumps({"@type": "TestSchema", "time_": "9:00:00", "id": "foo_item"}),
         )
         assert status == 201
 
-        resp, status = await requester(
-            "GET", "/db/guillotina/foo_item", data=json.dumps({"@type": "TestSchema", "time_": "16:40:09"})
-        )
+        resp, status = await requester("GET", "/db/guillotina/foo_item")
         assert status == 200
-        assert resp["time"] == "16:40:09"
+        assert resp["time_"] == "9:00:00"
 
         resp, status = await requester(
-            "POST", "/db/guillotina/", data=json.dumps({"@type": "TestSchema", "time_": "16:40:099", "id": "foo_item"})
+            "POST",
+            "/db/guillotina/",
+            data=json.dumps({"@type": "TestSchema", "time_": "9:70:00"}),
         )
         assert status == 412
 
         resp, status = await requester(
-            "POST", "/db/guillotina/", data=json.dumps({"@type": "TestSchema", "time_": 3600, "id": "foo_item"})
+            "POST",
+            "/db/guillotina/",
+            data=json.dumps({"@type": "TestSchema", "time_": "12:00:01"}),
         )
         assert status == 412
+        assert resp["deserialization_errors"][0]["message"] == "Value is too big"
+
+        resp, status = await requester(
+            "POST",
+            "/db/guillotina/",
+            data=json.dumps({"@type": "TestSchema", "time_": "8:59:59"}),
+        )
+        assert status == 412
+        assert resp["deserialization_errors"][0]["message"] == "Value is too small"
+        resp, status = await requester(
+            "POST",
+            "/db/guillotina/",
+            data=json.dumps({"@type": "TestSchema", "time_": 3600}),
+        )
+        assert status == 412
+
 
 async def test_patch_field_validation(container_requester):
     async with container_requester as requester: