+++ guillotina/tests/test_serialize.py
@@ -270,6 +270,12 @@ async def test_deserialize_time(dummy_guillotina):
     assert converted.minute == now.minute
 
 
+async def test_deserialize_time(dummy_guillotina):
+    now = time.fromisoformat("10:00:00")
+    converted = schema_compatible(now, ITestSchema["time"])
+    assert converted.minute == now.minute
+
+
 async def test_check_permission_deserialize_content(dummy_request):
     login()
     content = create_content()